+++ src/models/graphNode.ts
